split inspector into modules
